# Conflicts:
#	backend/team6/.gitignore
#	backend/team6/src/main/java/programmers/team6/Team6Application.java
#	backend/team6/src/main/java/programmers/team6/domain/auth/config/JwtAuthenticationFilter.java
#	backend/team6/src/main/java/programmers/team6/domain/auth/controller/AuthController.java
#	backend/team6/src/main/java/programmers/team6/domain/auth/dto/request/MemberSignUpRequest.java
#	backend/team6/src/main/java/programmers/team6/domain/auth/service/AuthService.java
#	backend/team6/src/main/java/programmers/team6/domain/auth/service/JwtService.java
#	backend/team6/src/main/java/programmers/team6/domain/auth/token/JwtTokenProvider.java
#	backend/team6/src/main/java/programmers/team6/domain/auth/util/JwtUtils.java
#	backend/team6/src/main/java/programmers/team6/domain/member/controller/MemberController.java
#	backend/team6/src/main/java/programmers/team6/domain/member/repository/MemberRepository.java
#	backend/team6/src/main/java/programmers/team6/domain/member/util/mapper/MemberMapper.java
#	backend/team6/src/main/java/programmers/team6/global/config/SecurityConfig.java
#	backend/team6/src/test/java/programmers/team6/domain/auth/service/AuthServiceTests.java

+++ backend/team6/src/main/java/programmers/team6/domain/vacation/dto/VacationMonthlyStatisticsResponse.java
@@ -1,21 +1,22 @@
 package programmers.team6.domain.vacation.dto;
 
 public record VacationMonthlyStatisticsResponse(
-        Long memberId,
-        String userName,
-        Long totalCount,
-        Long usedCount,
-        Long remainCount,
-        Long january,
-        Long february,
-        Long march,
-        Long april,
-        Long may,
-        Long june,
-        Long july,
-        Long august,
-        Long september,
-        Long october,
-        Long november,
-        Long december
-) {}
+	Long memberId,
+	String userName,
+	double totalCount,
+	double usedCount,
+	double remainCount,
+	double january,
+	double february,
+	double march,
+	double april,
+	double may,
+	double june,
+	double july,
+	double august,
+	double september,
+	double october,
+	double november,
+	double december
+) {
+}
